feat: refactoring and preparation for release
