+++ src-demo/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <base href="./" />
-    <title>CodeMirror6 - Block Indentation</title>
+    <title>Block Editor for CodeMirror 6</title>
   </head>
   <body>
     <div>
@@ -12,7 +12,7 @@
 
       <a
         style="float: right"
-        href="https://github.com/marcoklein/codemirror6-block-indentation"
+        href="https://github.com/marcoklein/codemirror-block-editor"
       >
         <img
           alt="View on GitHub"
