+++ src/index.ts
@@ -1,0 +1,6 @@
+export { blockLevelExtension } from "./extension/extension";
+export { blockLevelListenerFacet } from "./extension/block-level-listener";
+export {
+  BlockLevelChangeEffectSpec,
+  blockLevelChangeEffect,
+} from "./extension/set-block-level-effect";
